backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
openai>=1.6.0
yt-dlp>=2024.1.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
numpy>=1.24.0
